refactor: Standalone client version compare function
